api/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
basicauth==0.4.1
click==7.1.2
colorama==0.4.4
coverage==5.5
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.13.0
toml==0.10.2
Werkzeug==1.0.1